refactor: implement cross-compiling abstraction (#8)
