+++ src/logbypass/cpu.cc
@@ -1,6 +1,6 @@
 #include "cpu.h"
 #include "../logger.h"
-#include "uv.h"
+#include "../platform/platform.h"
 
 #define INIT_CPU_PERIOD(period)                                                \
   static double *cpu_##period = new double[period];                            \
@@ -41,73 +41,6 @@ double cpu_now = 0.0;
 INIT_CPU_PERIOD(15);
 INIT_CPU_PERIOD(30);
 INIT_CPU_PERIOD(60);
-
-#ifndef _WIN32
-double GetNowCpuUsage() {
-  static uint64_t last_time = 0;
-  static clock_t last_cpu_usage = 0;
-
-  // first time
-  if (last_time == 0 || last_cpu_usage == 0) {
-    last_time = uv_hrtime();
-    last_cpu_usage = clock();
-    return -1;
-  }
-
-  // calculate cpu usage
-  double cpu_now_ = 100 * (clock() - last_cpu_usage) / CLOCKS_PER_SEC /
-                    ((uv_hrtime() - last_time) / 10e8);
-
-  // update time & cpu usage
-  last_time = uv_hrtime();
-  last_cpu_usage = clock();
-
-  return cpu_now_;
-}
-#else
-#define GET_CURRENT_TIME(curtime)                                              \
-  HANDLE hProcess = OpenProcess(PROCESS_QUERY_INFORMATION, FALSE, getpid());   \
-  BOOL bRetCode = FALSE;                                                       \
-  FILETIME CreateTime, ExitTime, KernelTime, UserTime;                         \
-  LARGE_INTEGER lgKernelTime;                                                  \
-  LARGE_INTEGER lgUserTime;                                                    \
-  LARGE_INTEGER lgCurTimeTmp;                                                  \
-  bRetCode = GetProcessTimes(hProcess, &CreateTime, &ExitTime, &KernelTime,    \
-                             &UserTime);                                       \
-  if (!bRetCode) {                                                             \
-    return -1;                                                                 \
-  }                                                                            \
-  lgKernelTime.HighPart = KernelTime.dwHighDateTime;                           \
-  lgKernelTime.LowPart = KernelTime.dwLowDateTime;                             \
-  lgUserTime.HighPart = UserTime.dwHighDateTime;                               \
-  lgUserTime.LowPart = UserTime.dwLowDateTime;                                 \
-  lgCurTimeTmp.QuadPart =                                                      \
-      (lgKernelTime.QuadPart + lgUserTime.QuadPart) / 10000;                   \
-  curtime = lgCurTimeTmp;
-
-double GetNowCpuUsage() {
-  static LARGE_INTEGER g_slgProcessTimeOld = {0};
-  static uint64_t last_time = 0;
-
-  // first time
-  if (last_time == 0) {
-    GET_CURRENT_TIME(g_slgProcessTimeOld);
-    last_time = uv_hrtime();
-    return -1;
-  }
-
-  LARGE_INTEGER lgCurTime = {0};
-  GET_CURRENT_TIME(lgCurTime);
-  double cpu_now_ = ((lgCurTime.QuadPart - g_slgProcessTimeOld.QuadPart) * 100 /
-                     ((uv_hrtime() - last_time) / 10e5));
-
-  // update time & cpu usage
-  g_slgProcessTimeOld = lgCurTime;
-  last_time = uv_hrtime();
-
-  return cpu_now_;
-}
-#endif
 
 void SetNowCpuUsage() {
   double cpu_now_ = GetNowCpuUsage();
