+++ src/logbypass/log.cc
@@ -18,7 +18,7 @@ static void CreateUvThread(void *data) {
   uint64_t last_loop_time = uv_hrtime();
   while (1) {
     // sleep 1s for releasing cpu
-    SleepSeconds(1);
+    Sleep(1);
 
     // set now cpu usage
     SetNowCpuUsage();
