+++ src/utils.cc
@@ -1,17 +1,5 @@
-#ifndef _WIN32
-#include <unistd.h>
-#else
-#include <windows.h>
-#endif
-
-#include "utils.h"
+#include "platform/platform.h"
 
 namespace xprofiler {
-void SleepSeconds(int seconds) {
-#ifndef _WIN32
-  sleep(seconds);
-#else
-  Sleep(seconds * 1000);
-#endif
-}
+void Sleep(int seconds) { SleepCrossPlatform(seconds); }
 }; // namespace xprofiler
